Milestone data model, UI changes to show pill, no UI to add milestones yet
