+++ packages/shared/src/reducers/__tests__/backlogItemsReducerHelper.test.ts
@@ -5,7 +5,7 @@ import "jest";
 import produce from "immer";
 
 // code under test
-import { rebuildAllItems } from "../backlogItems/backlogItemsReducerHelper";
+import { rebuildAllItems } from "../backlogItems/helpers/backlogItemsReducerHelper";
 
 // interfaces/types
 import type { BacklogItemWithSource, SaveableBacklogItem } from "../backlogItems/backlogItemsReducerTypes";
