+++ packages/shared/src/types/apiModelTypes.ts
@@ -94,6 +94,19 @@ export type ApiBacklogItem = StandardItem &
     };
 
 export type ApiMilestone = StandardItem & {
+    backlogitemId: string;
+    milestoneId: string;
+    milestone: StandardItem & {
+        archived: string; // Y/N
+        id: string; // same as milestoneId
+        name: string;
+        projectId: string;
+        targetdate: ISODateString;
+    };
+    backlogitem: {
+        externalId: string | null;
+        friendlyId: string;
+    };
     name: string;
     backlogItemIds: string[];
 };
